Moved peer manager into its own file.
