+++ ntp-daemon/src/peer_manager.rs
@@ -1,0 +1,225 @@
+use crate::{
+    config::PeerConfig,
+    observer::ObservablePeerState,
+    peer::{MsgForSystem, ResetEpoch},
+};
+use ntp_proto::PeerSnapshot;
+
+#[derive(Debug, Clone, Copy)]
+pub enum PeerStatus {
+    /// This peer is demobilized, meaning we will not send further packets to it.
+    /// Demobilized peers are kept because our logic is built around using indices,
+    /// and removing a peer would mess up the indexing.
+    Demobilized,
+    /// We are waiting for the first snapshot from this peer _in the current reset epoch_.
+    /// This state is the initial state for all peers (when the system is spawned), and also
+    /// entered when the system performs a clock jump and forces all peers to reset, or when a peer
+    /// indicates that it is no longer fit for synchronization (e.g. root distance became too big)
+    ///
+    /// A peer can leave this state by either becoming demobilized, or by sending a snapshot that
+    /// is within the system's current reset epoch.
+    NoMeasurement,
+    /// This peer has sent snapshots taken in the current reset epoch. We store the most recent one
+    Measurement(PeerSnapshot),
+}
+
+#[derive(Debug, Clone, Copy)]
+pub struct PeerIndex {
+    pub index: usize,
+}
+
+#[derive(Debug, Default)]
+pub struct Peers {
+    peers: Box<[PeerStatus]>,
+    peer_configs: Vec<PeerConfig>,
+}
+
+impl Peers {
+    pub fn new(peer_configs: &[PeerConfig]) -> Self {
+        Self {
+            peers: vec![PeerStatus::NoMeasurement; peer_configs.len()].into(),
+            peer_configs: peer_configs.to_vec(),
+        }
+    }
+
+    #[cfg(test)]
+    pub fn from_statuslist(data: &[PeerStatus], peer_configs: &[PeerConfig]) -> Self {
+        Self {
+            peers: data.to_owned().into(),
+            peer_configs: peer_configs.to_vec(),
+        }
+    }
+
+    pub fn size(&self) -> usize {
+        self.peers.len()
+    }
+
+    pub fn observe(&self) -> impl Iterator<Item = ObservablePeerState> + '_ {
+        self.peers
+            .iter()
+            .zip(self.peer_configs.iter())
+            .map(|(status, config)| match status {
+                PeerStatus::Demobilized => ObservablePeerState::Nothing,
+                PeerStatus::NoMeasurement => ObservablePeerState::Nothing,
+                PeerStatus::Measurement(snapshot) => ObservablePeerState::Observable {
+                    statistics: snapshot.statistics,
+                    reachability: snapshot.reach,
+                    uptime: snapshot.time.elapsed(),
+                    poll_interval: snapshot.poll_interval.as_system_duration(),
+                    peer_id: snapshot.peer_id,
+                    address: config.addr.clone(),
+                },
+            })
+    }
+
+    pub fn valid_snapshots(&self) -> impl Iterator<Item = PeerSnapshot> + '_ {
+        self.peers
+            .iter()
+            .filter_map(|peer_status| match peer_status {
+                PeerStatus::Demobilized | PeerStatus::NoMeasurement => None,
+                PeerStatus::Measurement(snapshot) => Some(*snapshot),
+            })
+    }
+
+    pub fn update(&mut self, msg: MsgForSystem, current_reset_epoch: ResetEpoch) {
+        match msg {
+            MsgForSystem::MustDemobilize(index) => {
+                self.peers[index.index] = PeerStatus::Demobilized;
+            }
+            MsgForSystem::NewMeasurement(index, msg_reset_epoch, snapshot) => {
+                if current_reset_epoch == msg_reset_epoch {
+                    self.peers[index.index] = PeerStatus::Measurement(snapshot);
+                }
+            }
+            MsgForSystem::UpdatedSnapshot(index, msg_reset_epoch, snapshot) => {
+                if current_reset_epoch == msg_reset_epoch {
+                    self.peers[index.index] = PeerStatus::Measurement(snapshot);
+                }
+            }
+        }
+    }
+
+    pub fn reset_all(&mut self) {
+        for peer_status in self.peers.iter_mut() {
+            use PeerStatus::*;
+
+            *peer_status = match peer_status {
+                Demobilized => Demobilized,
+                Measurement(_) | NoMeasurement => NoMeasurement,
+            };
+        }
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use ntp_proto::{peer_snapshot, NtpDuration, NtpInstant, PeerStatistics};
+
+    use crate::config::PeerHostMode;
+
+    use super::*;
+
+    #[test]
+    fn test_peers() {
+        let base = NtpInstant::now();
+        let prev_epoch = ResetEpoch::default();
+        let epoch = prev_epoch.inc();
+        let mut peers = Peers::new(&test_peer_configs(4));
+        assert_eq!(peers.valid_snapshots().count(), 0);
+
+        peers.update(
+            MsgForSystem::NewMeasurement(
+                PeerIndex { index: 0 },
+                prev_epoch,
+                peer_snapshot(
+                    PeerStatistics {
+                        delay: NtpDuration::from_seconds(0.1),
+                        offset: NtpDuration::from_seconds(0.),
+                        dispersion: NtpDuration::from_seconds(0.05),
+                        jitter: 0.05,
+                    },
+                    base,
+                    NtpDuration::from_seconds(0.1),
+                    NtpDuration::from_seconds(0.05),
+                ),
+            ),
+            epoch,
+        );
+        assert_eq!(peers.valid_snapshots().count(), 0);
+
+        peers.update(
+            MsgForSystem::NewMeasurement(
+                PeerIndex { index: 0 },
+                epoch,
+                peer_snapshot(
+                    PeerStatistics {
+                        delay: NtpDuration::from_seconds(0.1),
+                        offset: NtpDuration::from_seconds(0.),
+                        dispersion: NtpDuration::from_seconds(0.05),
+                        jitter: 0.05,
+                    },
+                    base,
+                    NtpDuration::from_seconds(1.0),
+                    NtpDuration::from_seconds(2.0),
+                ),
+            ),
+            epoch,
+        );
+        assert_eq!(peers.valid_snapshots().count(), 1);
+
+        peers.update(
+            MsgForSystem::NewMeasurement(
+                PeerIndex { index: 0 },
+                epoch,
+                peer_snapshot(
+                    PeerStatistics {
+                        delay: NtpDuration::from_seconds(0.1),
+                        offset: NtpDuration::from_seconds(0.),
+                        dispersion: NtpDuration::from_seconds(0.05),
+                        jitter: 0.05,
+                    },
+                    base,
+                    NtpDuration::from_seconds(0.1),
+                    NtpDuration::from_seconds(0.05),
+                ),
+            ),
+            epoch,
+        );
+        assert_eq!(peers.valid_snapshots().count(), 1);
+
+        peers.update(
+            MsgForSystem::UpdatedSnapshot(
+                PeerIndex { index: 1 },
+                epoch,
+                peer_snapshot(
+                    PeerStatistics {
+                        delay: NtpDuration::from_seconds(0.1),
+                        offset: NtpDuration::from_seconds(0.),
+                        dispersion: NtpDuration::from_seconds(0.05),
+                        jitter: 0.05,
+                    },
+                    base,
+                    NtpDuration::from_seconds(0.1),
+                    NtpDuration::from_seconds(0.05),
+                ),
+            ),
+            epoch,
+        );
+        assert_eq!(peers.valid_snapshots().count(), 2);
+
+        peers.update(MsgForSystem::MustDemobilize(PeerIndex { index: 1 }), epoch);
+        assert_eq!(peers.valid_snapshots().count(), 1);
+
+        peers.reset_all();
+        assert_eq!(peers.valid_snapshots().count(), 0);
+    }
+
+    fn test_peer_configs(n: usize) -> Vec<PeerConfig> {
+        (0..n)
+            .map(|i| PeerConfig {
+                addr: format!("127.0.0.{i}:123"),
+                mode: PeerHostMode::Server,
+            })
+            .collect()
+    }
+}
